Enable zstandard supercompression for generated ktx2 files
This uses a more size-optimized basis_universal encoder bundle by
removing support for jpeg, png, exr etc loading which is not used in the
encoder anyway. This allows the inclusion of zstandard compression for
only 70kb more bundle size.

End result should be that cached ktx2 files are only about 1/3 to 1/4
the size as previously.