Scale video frames for spritesheet generation to fit maximum size.
This will actually try and get the 'default' scale of an effect that is
in use and request a spritesheet with that scaling as the set minimum.
Spritesheets are generated by first trying full size frames, 1/2, 1/4,
1/8... size frames until the minimum is met, bailing if the spritesheet
still exeeds the set maximum size. That way, spritesheets are gauranteed
to preserve as much detail as possible while still being able to be
generated at all.

Once such a scaled spritesheet is generated, it is added to the cache as
normal, but with its scaling factor also persisted.

Requests for compiled spritesheets that do not satisfy that minimum
scaling limit are not fullfilled, falling back to regular video based
effects instead.