Report spritesheet currentTime with ms accuracy to match video files
Essentially, the video element reports duration, currentTime etc rounded
with ms accuracy but spritesheets were reported with raw frametime accuracy.
This caused issues when calculating the total time for loop handling
based on the video file but switched to spritesheets for optimization.
We checked in the loop handler if the currentTime matched the calculated
(rounded) end time, which in certain cases was never the case because of
this rounding, sometimes missing 0.4ms or something.