Various minor fixes
Discovered errors: Effects with only a text would not play because
sprite.anchor was undefined and optional chaining was missing.

xRay setting was not working because the property for that setting was
`isVisionMaskingEnabled`, but the setter was called `isVisionMaskEnabled`,
leading to that flag never being set correctly.

Spritesheets under certain circumstances threw errors in the cleanup
code. This had no effet other than console error messages being logged

If above ui effects layer was disabled, after at least one effect was
played on the regular screen space layer and it ended, the layer itself
was marked as not rendereable, leading to future effects not being
displayed anymore