Fix calculation of dimensions for canvas tiles
Reason for this PR is that effects being attached to tiles and having
`scaleToObject` set were invisible. Underlying reason for that is that
the Tile object `width` and `height` values are 0 when not also
displaying the token border. This is most likely a Foundry V12 change
that went unnoticed.

As a fix, we can query the Tiles bounds instead. I made sure that this
produces the correcdt results for tiles, tokens, templates, plain
coordinates and crosshairs.

I also verified that shapes and screen space (above) ui is not broken.